TMDB API and Cinemeta working
Programmed methods for trying getting TMDB metadata (which Letterboxd uses), including id.

In case of failure with TMDB, the app tries the Cinemeta Addon, which gets us roughly the same info (excluding if it's adult content or not).

Also, whenever we succeed, we add a sub entry in Stremio starting with LB- that lets us know the app worked and with wich source.

Cleaned up the codebase and updated the README.

This resolves #2